Added a script for materialized views refreshing.
For the given MATVIEWS_DBNAME the script discovers all the
materialized views and their indexes, drops the indexes concurrently,
refreshes the materialized views in their dependencies depth order,
and creates their indexes concurrently. The script is primarily
designed to run as a cronjob at periods of inactivity. Compatible with
PostgreSQL >=9.3.